requirements.txt
Flask>=2.0
pandas
numpy
scikit-learn
joblib
requests
httpx
dnspython
python-whois
cryptography
httpx
tldextract
python-dotenv
